launch.sh: use general playbook for config and test
Calling the playbooks one by one directly in the CI is not easy to
maintain : For adding a playbook, one has to add it in the Ansible
repository and call it here in the CI.

To avoid this problem, the script now call `ci_configure.yaml` and
`ci_test.yml`. These playbooks will call the necessary ones directly in
Ansible.

Signed-off-by: Erwann Roussy <erwann.roussy@savoirfairelinux.com>