report-generator: make latency test results optional
Asciidoc format allow an include to be optional and that's exactly what
is needed for the latency test part:

If cukinia tests fail, the latency part will not launched (this is done
to reduce the CI time in case of failure). In that case, the latency
results file doesn't exist. Making it optionnal is way better than just
touching the file in the ci script.

Signed-off-by: Erwann Roussy <erwann.roussy@savoirfairelinux.com>