launch.sh: add test report generation for VM part
The CI test part is a bit tricky, it now contains three parts :
- cukinia tests on physical machines
- cukinia tests on VM
- lantency tests

These three parts are interconnected :
- If test on physical machines fails, we generate the test report for
  this part but don't launch the other tests
- If VM test fails, we generate the test report for all cukinia tests but
  don't launch latency tests
- If both path are ok, we launch latency test and generate the report
  with these three parts.

In order to achive that, all test step generate their own adoc file.
The pdf file is generated in the last step using all adoc file available
( that depends on which test step was launch or not )

My problem is that the VM test part must be integrated with the physical
machine test part in order for the compliance matrix to work.
To achieve that, I regenerate all the adoc file in the VM section,
including cukinia test file for physical machine and VM.

TLDR:
- launch_system_tests generate an adoc file at the end. This adoc file
  contains cukinia tests results file for hypervisors.
- launch_vm_tests regenerates the same adoc file but including the vm
  cukinia test result file.

Signed-off-by: Erwann Roussy <erwann.roussy@savoirfairelinux.com>