launch.sh: use test-report-pdf
Move the report generation to test-report-pdf.

This includes to do a trick with the latency results.
Because they have to be printed at the end of the report, the adoc file
containing the part about the latency results is copied in
test-report-pdf under the name "notes.adoc".
By doing so, this file will be detected by test-report-pdf and put at
the end.

Signed-off-by: Erwann Roussy <erwann.roussy@savoirfairelinux.com>