launch.sh: correct vm tests grep.
A grep command is launched after the VM tests to know if any test is
failed. However, this command is launched over the directory containing
all cukinia results, including the one from the hypervisors.

One of the possible test failure in debian CI is a kernel backtrace that
we want to ignore (see 5a897ea).
The ignore was done properly in the hypervisor test part, but not in the
vm test part.
This commit correct the grep command to ignore the backtrace test
failure.

Signed-off-by: Erwann Roussy <erwann.roussy@savoirfairelinux.com>