launch.sh: rename cluster playbooks
As discussed with Florent in
https://github.com/seapath/ansible/pull/585#issuecomment-2306561089
we should stop using standard playbooks on VMs.
This commit start to do the job by calling directly the prerequisdebian
on the VM, thus avoiding the network playbook to be applied.
Some required setup is done in a new playbook called ci_prepare_VMs.

Also, cluster_setup_hardened_debian is now renamed to
seapath_setup_hardened_debian

Signed-off-by: Erwann Roussy <erwann.roussy@savoirfairelinux.com>