launch-yocto: analyse only SV stream 0 for latency test
Currently, no matter the number of recorded and analysed streams,
sv_timestamp_analysis only includes SV stream 0 results
in the generated .adoc report.

Due to this limitation, and because sv_timestamp_analysis is for now
unable to efficiently analyse a large amount of data, limit for now the
analysis to SV stream 0.
This will allow in future commits to run longer latency tests with
the currently available tools and hardware.

Signed-off-by: Elinor Montmasson <elinor.montmasson@savoirfairelinux.com>