launch-yocto: reduce the maximum latency threshold for latency tests
Currently, the maximum latency test threshold was 2000 µs.
However, the measured maximum latency rarely goes over 200 µs.
Reduce the threshold to 250 µs to improve the relevance of the test.
It also improves the relevance of the latency result histogram in the
report, as the Y axis will be better adapted to the results.

Signed-off-by: Elinor Montmasson <elinor.montmasson@savoirfairelinux.com>