launch-yocto.sh: move latency test result copy
If the latency test fails, the check will automatically stop the latency
function.
In that case, the copy commands at the end of the function will not be
executed, and thus not displayed in the test report.

This commit changes the location of the copy commands, so the histogram
are visible on the report, even if the latency tests fail.

Signed-off-by: Erwann Roussy <erwann.roussy@savoirfairelinux.com>