launch-yocto: raise the AAEON latency test threshold to 1s
The AAEON board is currently not configured with realtime capabilities.
Recently, it happened that the previous threshold (10ms) wasn't enough,
so the CI would fail due to this test.
Until work is done to configure realtime on the AAEON board, raise the
latency threshold to 1s to unblock the CI.
The latency threshold should be lowered as soon as possible.

Signed-off-by: Elinor Montmasson <elinor.montmasson@savoirfairelinux.com>