launch-yocto.sh: add svtrace integration
Currently, latency tests are done using dedicated playbooks in SEAPATH
Ansible directory.
Developed in 2024, SEAPATH svtrace is a tool used to perform IEC61850
latency tests. This repository is used to maintain a common foundation
to run latency tests on SEAPATH. Plus, the build of sv_timestamp_logger
is now done by svtrace, with a detection of the target architecture.

This commit makes the modification to replace the latency tests run in
Ansible playbooks, by svtrace-ansible. The results are the same
(generated report).

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>